Add method for getting the routing of solution
This is a list of lists. Each element represents the path of pickup and
delivery tour that every vehicle does.

Example:

```
routing = [
    ['a', 'b', 'c'], ['d', 'e'],
    ['f', 'g'], ['h', 'i', 'j'],
]
```

We used two vehicles. For instance, the first vehicle visited three
suppliers with the following order ('a', 'b', 'c'), then it returned
to cross dock and it delivered to 'd', and 'e'.